Modify 1-35,42,43, Add 1-36,37
